move all methods from controller to helper class without any changes in individual methods
